requirements.txt
ArabicOcr==1.1.6
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
CacheControl==0.12.10
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.4
colorama==0.4.4
crypto==1.4.1
distlib==0.3.3
Django==3.2.9
django-cors-headers==3.11.0
django-environ==0.8.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
docopt==0.6.2
easyocr==1.4.1
filelock==3.3.2
filestack-python==3.5.0
firebase-admin==5.1.0
Flask==2.0.3
gcloud==0.18.3
google-api-core==2.2.2
google-api-python-client==2.32.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-cloud-core==2.2.1
google-cloud-firestore==2.3.4
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
grpcio==1.42.0
grpcio-status==1.42.0
httplib2==0.20.2
idna==3.3
imageio==2.9.0
itsdangerous==2.1.0
Jinja2==3.0.3
joblib==1.1.0
jws==0.1.3
MarkupSafe==2.1.0
msgpack==1.0.3
Naked==0.1.31
networkx==2.7.1
numpy==1.22.3
oauth2client==4.1.3
opencv-python-headless==4.5.5.62
packaging==21.3
pandas==1.4.1
pexpect==4.8.0
Pillow==8.2.0
pipreqs==0.4.11
platformdirs==2.4.0
proto-plus==1.19.8
protobuf==3.19.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.11.0
PyJWT==1.7.1
pyngrok==5.1.0
pyocr==0.8.1
pyparsing==3.0.6
Pyrebase4==4.5.0
pytesseract==0.3.9
python-bidi==0.4.2
python-dateutil==2.8.2
python-jwt==2.0.1
pytz==2018.9
PyWavelets==1.2.0
PyYAML==6.0
requests==2.26.0
requests-toolbelt==0.9.1
rsa==4.7.2
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
shellescape==3.8.1
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
threadpoolctl==3.1.0
tifffile==2022.2.9
torch==1.10.2
torchvision==0.11.3
tqdm==4.63.0
trafaret==2.0.2
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.7
virtualenv==20.10.0
Werkzeug==2.0.3
yarg==0.1.9
